Correção no PedidoService e PedidoResponseDTO. Criaçaõ dos méteodos restantes do CRUD de Prato

+++ src/main/java/br/com/ifpb/hibernate_restaurant_order_system/dto/pedido/PedidoResponseDTO.java
@@ -3,6 +3,8 @@ package br.com.ifpb.hibernate_restaurant_order_system.dto.pedido;
 import br.com.ifpb.hibernate_restaurant_order_system.dto.prato.PratoResponseDTO;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,7 +14,7 @@ public class PedidoResponseDTO {
     // Atributos necessários para responder sobre um Pedido
     private Long id;
     private String data;
-    private Double valorTotal;
+    private BigDecimal valorTotal;
     private Long clienteId;    
     private List<PratoResponseDTO> pratos;
