+++ src/main/java/br/com/ifpb/hibernate_restaurant_order_system/service/PedidoService.java
@@ -97,7 +97,7 @@ public class PedidoService {
         return new PedidoResponseDTO(
                 pedido.getId(),
                 pedido.getData().toString(),
-                pedido.getValorTotal() == null ? null : pedido.getValorTotal().doubleValue(),
+                pedido.getValorTotal() == null ? null : pedido.getValorTotal(),
                 pedido.getCliente().getId(),
                 pratosDto
         );
